Fix installation process
The generate.js file required the app/content/texts/ directory to exist.
The new .gitignore and .gitkeep files will ensure the directory is
there. The .gitignore file will make sure that no biblical texts in that
directory get added to the project.

I also updated the README file to reflect the correct location of the
generate.js file.