Fixed lemma popup disappearing.
If the lemma popup was already displayed and a click happened elsewhere
on the document then the InfoWindow plugin would hide the lemma popup
which is proper behavior.

If the click happened on another lemma then the lemma popup would be
updated with new lemma information and moved to the new location by
the lemma plugin click handler. Then the click would bubble up to the
document click handler created InfoWindow and hide the updated lemma
popup.  The result would be the appearance that the lemma popups had
quit working when in fact they were being displayed and immediately
hidden.

To correct this problem, the lemma popup is now hidden when ever it is
already visible and then redisplayed with the new lemma information.
This fixes the problem because the hide() function disables the document
click handler before the lemma click event bubbles up.