normalize whitespace in CSS files
If there was a rhyme or reason to the indent levels is the CSS files I
couldn't make out what it was. It looks like it might have been leftover
from a templating system or nested structure.

Otherwise this just normalizes coding style to remove the single-line
entries that are a pain to edit, esp in an editor with line-wrap turned
off.

Some more cleanup could be done to properly comment sections and fix the
inconsistent breaking between related and unrelated items.