remove obsolete webfont formats, leave only WOFF
The switch to jQuery 2 already rules out all of the browsers that would
be picked up by the alternate font formats. EOT is only needed for IE8,
but 9 is required for jQuery 2. The TTF and SVG formats only picked up
older mobile platforms, but all the ones supported by jQuery (and even
2 versions older) support WOFF.