fix fall-back on custom font declaration
Both ways work in the end, but the fall-back happening here is not from one
font family to another but from a possible local installation of the same font
to using a webfont version from the app. The CSS handling this fall-back
should be the font-face declaration, not the family selector.