fix CSS parse errors and work around IE bug
A couple of these lines just had bad syntax (missing colons). The rest
of this is white spaces fixes that technically should not be necessary
as the syntax is correct per spec, but IE 7-9 has an issue parsing it.
As IE9 is supported by jQuery 2 and hence technically by this app, I've
accommodated it by giving it the value it can read.