attach font changer only to font radio buttons
The function to set the font family was being attached to all input
elements in the entire document which included the font-size selector.
The result was that changing the font size would strip the font family
class off the body element and replace it with nonsense like
`class="config-font-family-18"`. This meant every time you adjusted the
size, your font pref disappeared and hand to be reset.

Fixes #20