fix regression in CSS cleanup from 9884e1a
CSSTidy apparently cannot handle vendor specific prefixes in values. It
handles them alright in attributes but choked on values. They apparently
get interpreted as a negative number, the string converted to an int,
and the result is "0". The result was a bogus 1s delay being applied in
lieu of a 3D transition. Webkit opted to apply this to all shape
changes, including window resizes.