set min-width for two columns to match mobile
With the current fixed element width system an argument could be made
for increasing this to 680 where everything shows up properly, but the
20 px difference between the mobile layout cutoff and the two column cut
off makes no sense. If nothing else it just makes a broken interface
where one of the major buttons does nothing. If these are to be at
different places, the add column button would also need to be hidden
separately.